db-connect-test/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
werkzeug==2.2.3
mysql-connector-python>=8.0.33
